test-runner: Scale more aggressively
This commit updates the test-runner autoscaler configurations such that
it scales out more aggressively.

This is necessary because the runner pods take longer to start up due
to the large size of the new pod image with the cached workflow Docker
images and the autoscaler does not scale out until the pods are
running and assigned a job, which can take minutes.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>