test-runner: Use private ECR image
This commit updates the test-runner to use an image from the private
ECR repository so that any ECR image pull traffic can be routed
internally through a VPC endpoint without going through the NAT
gateway.

Note that AWS does not allow creating a VPC endpoint for the public ECR
repositories.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>