zephyr-alpha: Add VPC endpoints for S3 and ECR services
This commit adds the VPC endpoints for S3 and ECR services so that the
container image pull requests are served internally without going
through the VPC NAT gateway.

Note that the ECR image layers are saved in the S3 storage and the S3
VPC endpoint must also be added to serve the image layer contents.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>