zephyr-alpha: Set EFS storage class uid and gid to 1000
This commit updates the EFS storage class definition to set the uid and
gid for accessing the EFS filesystem to 1000.

This ensures that the EFS volumes and its contents are assigned the uid
and gid of 1000, which is often the id for the default non-root user,
and not assigned an arbitrary gid.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>