zephyr-alpha: Update GitHub repository URL
This commit updates the GitHub repository URL specified in the cluster
configurations to point to the infrastructure-private repository.

Note that this is a mostly cosmetic change since this field is not
directly parsed by anything at this time; however, it may be used in
the future to filter by different cluster deployments.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>