test-runner: Update for custom AMI
This commit updates the test-runner to use the custom
`zephyr-runner-node` AMI, which includes the cached CI Docker images
and the repositories.

The dual "runner" and "dind" container scheme has also been reverted
to use the single "DinD in runner" container scheme because the Actions
Runner Controller sets the pod `restartPolicy` to `Never` and this
causes the pod to keep running if only the "dind" container terminates
for whatever reason.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>