test-runner: Use TotalNumberOfQueuedAndInProgressWorkflowRuns
This commit updates the test-runner to use the
"TotalNumberOfQueuedAndInProgressWorkflowRuns" autoscaling metric,
which scales the runner pods based on the number of queued and running
workflow jobs.

While the "PercentageRunnersBusy" is fairly stable, it launches too
many excess runner pods (as high as 1.5x of the number of required pods
in practice), resulting in a much higher operating cost than
anticipated.

As for the webhook-based scaling, it is simply too unstable for
production use.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>