kubernetes: Remove runner-repo-cache
This commit removes the Kubernetes EFS-based runner repository cache
deployment configurations because it has been superseded by the AMI-
based repository cache.

While the EFS-based repository cache worked reasonably well, it was
way too expensive to be practical especially with the amount of
provisioned I/O performance necessary to make the CI workflows complete
in a reasonable amount of time.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>