zephyr-aws-blueprints: Add sc1 storage class
This commit adds the 'sc1' Kubernetes storage class, which uses the AWS
SC1 cold HDD storage as backend.

The 'sc1' HDD storage type is significantly cheaper than the 'gp2' and
'gp3' SSD storage types and intended to be used for storing a large
amount of data that is less frequently accessed.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>