zephyr-aws-blueprints: Use zephyr-runner AMI 2024-02-04
This commit updates the zephyr-runner deployment to use the AMI
2024-02-04, which includes the CI Docker image v0.26.7.

Note that this only updates the x86-64 AMI because the AArch64 AMI is
not built due to the missing Zephyr SDK release for the architecture.

Also note that this uses the AMI for zephyr-runner v1 because v2 is
still work-in-progress.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>