zephyr-runner-v2: Shorten runner scale set name
This commit shortens the names of the zephyr-runner v2 runner scale sets
because they are directly used to name Kubernetes pods, and Kubernetes
limits the maximum length of pod name to 63 characters, which can be
exceeded with the extra tags appended by the runner scale set
controller.

Note that this leads to the GitHub runner labels to be mapped to the
shortened names as well because there currently exists no option to
specify a separate runner label from the runner scale set name -- this,
however, is not a problem because the zephyr-runner v2 deployment is
intended to be referenced in GitHub Actions workflows by the runner node
group name, which is not shortened.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>