zephyr-runner-v2: test-runner-cnx: Set cloud information as envvars
This commit adds the runner environment variables that provide the
underlying cloud service information for debugging and cloud-specific
workflow handling purposes.

As part of this, it adds the pod templates for instantiating the
workflow containers, which is necessary in order to make the cloud
service information environment variables available inside containers.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>