zephyr-runner-v2: Increase ACTIONS_RUNNER_PREPARE_JOB_TIMEOUT_SECONDS
This commit increases the ACTIONS_RUNNER_PREPARE_JOB_TIMEOUT_SECONDS
from the default value of 600 seconds to 3600 seconds because the CI
Docker images are very large and pulling them for the first time can
take a long time.

Note that the CI Docker images are not cached in the Centrinix
deployment because the runner nodes are non-ephemeral and therefore only
need to download the CI Docker images once after provisioning.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>