zephyr-runner-v2: cnx: Add node selector for Magnum node group role
The OpenStack Magnum node groups for the runner nodes now have the role
value of 'runner'.

This commit adds the node selector that selects the Magnum node groups
with the runner role to the zephyr-runner v2 pods so that they are not
scheduled onto any other node types (e.g. cache).

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>