zephyr-runner-v2: cnx: Add rawfile-localpv Helm chart configuration
This commit adds the Helm chart configuration for OpenEBS
rawfile-localpv, which provides sparsefile-based node local storage.

While OpenEBS hostPath local PV works, it uses the node filesystem and,
when a provisioned volume contains millions of files, it takes a very
long time to clean up.

rawfile-localpv creates a "disk image" on a node and mounts it into the
pods using loopback devices -- with this, as far as the node filesystem
is concerned, only one disk image file needs to be deleted as part of
the cleanup.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>