ci: cloud-monitor: Add stuck PV clean-up step
This commit adds a step to clean up stuck Kubernetes PVs.

Note that this is necessary because rawfile-localpv sometimes fails to
delete the backing image file thinking the file is in use when it in
fact is not -- manually deleting the backing file nudges the
rawfile-localpv CSI driver to execute rest of the cleanup process.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>