terraform: Remove aws-zephyr-beta deployment
This commit removes the Terraform manifests for the `aws-zephyr-beta`
AWS EKS Kubernetes cluster, which is a backup EKS instance dedicated for
testing, because it is no longer active and there is no plan to
reinstate it in the future.

Any future Kubernetes testing should be done on the main `zephyr-alpha`
cluster or in a separate manually created EKS cluster.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>