zephyr-runner-v2: hzr: Set ARM64 Linux runner max count to 5
This commit sets the Hetzner ARM64 Linux runner count to 5, which
corresponds to the actual hardware capacity in the Hetzner cluster --
note that there is currently 1 RX220 server with 80 cores.

This prevents the GitHub from over-scheduling jobs to the Hetzner ARM64
runner scale set and effectively allows the remaining jobs to be
scheduled to the CNX ARM64 runner scale set.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>