zephyr-runner-v2: cnx: Set ARM64 Linux runner max count to 10
This commit sets the Centrinix ARM64 Linux runner count to 10, which
corresponds to the actual hardware capacity in the Centrinix cluster --
note that there is currently 1 Ampere Mt. Collins server with 160 cores.

This prevents the GitHub from over-scheduling jobs to the Centrinix
ARM64 runner scale set and effectively allows the remaining jobs to be
scheduled to the HZR ARM64 runner scale set.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>