sync_team_membership: Add maintainers and collaborators to contributors
Add all maintainers and collaborators listed in the Zephyr MAINTAINERS
file to the 'contributors' team.

This ensures that, when a collaborator is promoted to maintainer (i.e.
when a user is removed from the 'collaborators' team and added to the
'maintainers' team) or vice versa, the user is not removed from and
re-added to the zephyrproject-rtos organisation.

Note that this should be automatically done even without the
intervention of this script in the future because the documented process
requires that a user be a contributor prior to being promoted to a
collaborator or a maintainer.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>