sync_team_membership: Add maintainers to collaborators team
By definition, all maintainers are collaborators -- from Zephyr TSC
Project Roles documentation:

  A Maintainer is a Collaborator who is also responsible for knowing,
  directing and anticipating the needs of a given zephyr source code
  area.

This commit updates the team membership sync script to add all
maintainers to the 'collaborators' team such that the 'collaborators'
team is a superset of the 'maintainers' team.

Note that the same logic applies for collaborators and contributors,
which is already handled.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>