workflows: Add workflow for syncing module with upstream
Since some modules owned by the Zephyr organization are not created as a
Github fork maintaining them as a complete mirror requires manual
pushing. To fix this add a reusable workflow that can be configured for
each module to sync a specific branch in our version with upstream.
Also adds a sample usage of said workflow for the LVGL module.

Signed-off-by: Fabian Blatz <fabianblatz@gmail.com>