github: Add append-only team membership sync script
This commit adds the `maintainers` and `collaborators` team membership
sync script that only appends new users added to the `MAINTAINERS.yml`
file without processing removals.

The purpose of this is to let the previously active Zephyr maintainers
and collaborators who are no longer listed in the `MAINTAINERS.yml` file
keep their maintainer/collaborator-level permissions in case they want
to come back from time to time and help review and/or contribute.

This is meant to serve as a temporary solution until we have a more
concrete scheme for managing the "honourary maintainers/collaborators."

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>